+++ app/root.tsx
@@ -1,6 +1,5 @@
 import { Page } from "@navikt/ds-react";
 import { injectDecoratorClientSide } from "@navikt/nav-dekoratoren-moduler";
-import parse from "html-react-parser";
 import { useEffect } from "react";
 import {
   data,
@@ -9,7 +8,6 @@ import {
   Outlet,
   Scripts,
   ScrollRestoration,
-  useLoaderData,
   type HeadersArgs,
   type LoaderFunctionArgs,
 } from "react-router";
@@ -17,18 +15,10 @@ import type { Route } from "./+types/root";
 import "./app.css";
 import { env } from "./config/env.server";
 import { lagDekoratørHtmlFragmenter } from "./features/dektoratøren/htmlFragmenter";
-import { useDekoratørSpråk } from "./features/dektoratøren/useDektoratørSpråk";
-import { useInjectDecoratorScript } from "./features/dektoratøren/useInjectDecoratorScript";
 import { InternalServerError } from "./features/feilhåndtering/500";
 import { lagHeaders } from "./features/headers/headers.server";
-import { Analytics } from "./utils/analytics";
 import { fåApplikasjonsside } from "./utils/applikasjonssider";
-import {
-  definerTekster,
-  hentSpråkFraCookie,
-  oversett,
-  OversettelseProvider,
-} from "./utils/i18n";
+import { hentSpråkFraCookie } from "./utils/i18n";
 
 export async function loader({ request }: LoaderFunctionArgs) {
   const url = new URL(request.url);
@@ -58,41 +48,7 @@ export const headers = ({ loaderHeaders }: HeadersArgs) => {
 };
 
 export default function App() {
-  const { dekoratørHtml, språk, umamiWebsiteId, applikasjonsside } =
-    useLoaderData<typeof loader>();
-
-  const interntSpråk = useDekoratørSpråk(språk, applikasjonsside);
-  useInjectDecoratorScript(dekoratørHtml.DECORATOR_SCRIPTS);
-
-  return (
-    <html lang={oversett(interntSpråk, tekster.langTag)}>
-      <head>
-        <meta charSet="utf-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <Meta />
-        <Links />
-        {parse(dekoratørHtml.DECORATOR_HEAD_ASSETS, { trim: true })}
-        <Analytics umamiWebsiteId={umamiWebsiteId} />
-      </head>
-      <body className="flex flex-col min-h-screen">
-        {parse(dekoratørHtml.DECORATOR_HEADER, { trim: true })}
-        <OversettelseProvider språk={interntSpråk}>
-          <Page.Block
-            className="flex-1"
-            as="main"
-            id="maincontent"
-            width="xl"
-            gutters
-          >
-            <Outlet />
-          </Page.Block>
-        </OversettelseProvider>
-        {parse(dekoratørHtml.DECORATOR_FOOTER, { trim: true })}
-        <ScrollRestoration />
-        <Scripts />
-      </body>
-    </html>
-  );
+  return <Outlet />;
 }
 
 export function ErrorBoundary({ error }: Route.ErrorBoundaryProps) {
@@ -134,11 +90,3 @@ export function ErrorBoundary({ error }: Route.ErrorBoundaryProps) {
     </html>
   );
 }
-
-const tekster = definerTekster({
-  langTag: {
-    nb: "nb-NO",
-    en: "en-US",
-    nn: "nn-NO",
-  },
-});
